requirements.txt
pandas==1.4.1
numpy==1.22.3
jupyterlab==3.3.2
opentree==1.0.1
taxon_names_resolver==1.0.7
flatten_json==0.1.13
pydantic==1.9.0
python-dotenv==0.20.0